requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.14
dnspython==2.4.2
faiss-cpu==1.7.4
Flask==2.3.3
frozenlist==1.4.0
google-search-results==2.4.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.288
langsmith==0.0.36
loguru==0.7.2
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.6
numpy==1.25.2
oauthlib==3.2.2
openai==0.28.0
packaging==23.1
pinecone-client==2.2.4
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
PyMySQL==1.1.0
pypdf==3.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==2.0.20
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
tweepy==4.14.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
yarl==1.9.2
